+++ src/mo_def/syntax.ml
@@ -246,7 +246,13 @@ and dec' =
   | TypD of typ_id * typ_bind list * typ       (* type *)
   | ClassD of                                  (* class *)
       exp option * sort_pat * obj_sort * typ_id * typ_bind list * pat * typ option * id * dec_field list * function_context (* constructor context *)
+  | MixinD of pat * dec_field list             (* mixin *)
+  | IncludeD of id * exp * include_note (* mixin include *)
+and include_note' = { imports : import list; pat : pat; decs : dec_field list }
+and include_note = include_note' option ref
 
+and import = (import', Type.typ) Source.annotated_phrase
+and import' = pat * string * resolved_import ref
 
 (* Program (pre unit detection) *)
 
@@ -266,9 +272,6 @@ and req = bool Source.phrase
 
 (* Compilation units *)
 
-type import = (import', Type.typ) Source.annotated_phrase
-and import' = pat * string * resolved_import ref
-
 type comp_unit_body = (comp_unit_body', typ_note) Source.annotated_phrase
 and comp_unit_body' =
  | ProgU of dec list                         (* main programs *)
@@ -276,6 +279,7 @@ and comp_unit_body' =
  | ModuleU of id option * dec_field list     (* module library *)
  | ActorClassU of                            (* IC actor class, main or library *)
      persistence * exp option * sort_pat * typ_id * typ_bind list * pat * typ option * id * dec_field list
+ | MixinU of pat * dec_field list            (* Mixins *)
 
 type comp_unit = (comp_unit', prog_note) Source.annotated_phrase
 and comp_unit' = {