+++ src/pipeline/pipeline.ml
@@ -177,6 +177,7 @@ let async_cap_of_prog prog =
   match (CompUnit.comp_unit_of_prog false prog).it.body.it with
   | ActorClassU _ -> Async_cap.NullCap
   | ActorU _ -> Async_cap.initial_cap()
+  | MixinU _ -> Async_cap.initial_cap()
   | ModuleU _ -> assert false
   | ProgU _ ->
      if !Flags.compiled then
@@ -755,7 +756,12 @@ let run_files_and_stdin files =
 
 (* Desugaring *)
 
-let desugar_unit imports u name : Ir.prog =
+let desugar_unit imports u name : Ir.prog Diag.result =
+  match u.Source.it.Syntax.body.Source.it with
+  | Syntax.MixinU _ ->
+    let at = u.Source.it.Syntax.body.Source.at in
+    Diag.error at "M0225" "compile" "A mixin cannot be used as an entry point. It needs to be included in an actor (class)"
+  | _ ->
   phase "Desugaring" name;
   let open Lowering.Desugar in
   let prog_ir' : Ir.prog = link_declarations
@@ -764,7 +770,7 @@ let desugar_unit imports u name : Ir.prog =
   dump_ir Flags.dump_lowering prog_ir';
   if !Flags.check_ir
   then Check_ir.check_prog !Flags.verbose "Desugaring" prog_ir';
-  prog_ir'
+  Diag.return prog_ir'
 
 (* IR transforms *)
 
@@ -875,7 +881,8 @@ let rec compile_libs mode libs : Lowering.Desugar.import_declaration =
         Cons.session ~scope:filename (fun () ->
           match cub.it with
           | Syntax.ActorClassU _ ->
-            let wasm = compile_unit_to_wasm mode imports l in
+            (* Okay to "run" here, as `compile_unit_to_wasm` only fails on Syntax.MixinU *)
+            let wasm = Diag.run (compile_unit_to_wasm mode imports l) in
             Lowering.Desugar.import_compiled_class l wasm
           | _ ->
             Lowering.Desugar.import_unit l)
@@ -883,25 +890,27 @@ let rec compile_libs mode libs : Lowering.Desugar.import_declaration =
       go (imports @ new_imports) libs
   in go [] libs
 
-and compile_unit mode do_link imports u : Wasm_exts.CustomModule.extended_module =
+and compile_unit mode do_link imports u : Wasm_exts.CustomModule.extended_module Diag.result =
+  let open Diag.Syntax in
   let name = u.Source.note.Syntax.filename in
   Cons.session ~scope:name (fun () ->
-    let prog_ir = desugar_unit imports u name in
+    let* prog_ir = desugar_unit imports u name in
     let prog_ir = ir_passes mode prog_ir name in
     phase "Compiling" name;
     adjust_flags ();
     let rts = if do_link then Some (load_as_rts ()) else None in
-    if !Flags.enhanced_orthogonal_persistence then
+    Diag.return (if !Flags.enhanced_orthogonal_persistence then
       Codegen.Compile_enhanced.compile mode rts prog_ir
     else
-      Codegen.Compile_classical.compile mode rts prog_ir)
+      Codegen.Compile_classical.compile mode rts prog_ir))
 
-and compile_unit_to_wasm mode imports (u : Syntax.comp_unit) : string =
-  let wasm_mod = compile_unit mode true imports u in
+and compile_unit_to_wasm mode imports (u : Syntax.comp_unit) : string Diag.result =
+  let open Diag.Syntax in
+  let* wasm_mod = compile_unit mode true imports u in
   let (_source_map, wasm) = Wasm_exts.CustomModuleEncode.encode wasm_mod in
-  wasm
+  Diag.return wasm
 
-and compile_progs mode do_link libs progs : Wasm_exts.CustomModule.extended_module =
+and compile_progs mode do_link libs progs : Wasm_exts.CustomModule.extended_module Diag.result =
   let imports = compile_libs mode libs in
   let prog = CompUnit.combine_progs progs in
   let u = CompUnit.comp_unit_of_prog false prog in
@@ -911,7 +920,7 @@ let compile_files mode do_link files : compile_result =
   let open Diag.Syntax in
   let* libs, progs, senv = load_progs ~check_actors:true parse_file files initial_stat_env in
   let idl = Mo_idl.Mo_to_idl.prog (progs, senv) in
-  let ext_module = compile_progs mode do_link libs progs in
+  let* ext_module = compile_progs mode do_link libs progs in
   (* validate any stable type signature, as a sanity check *)
   let* () =
     match Wasm_exts.CustomModule.(ext_module.motoko.stable_types) with
@@ -937,18 +946,19 @@ let import_libs libs : Lowering.Desugar.import_declaration =
   List.concat_map Lowering.Desugar.import_unit libs
 
 let interpret_ir_progs libs progs =
+  let open Diag.Syntax in
   let prog = CompUnit.combine_progs progs in
   let name = prog.Source.note.Syntax.filename in
   let imports = import_libs libs in
   let u = CompUnit.comp_unit_of_prog false prog in
-  let prog_ir = desugar_unit imports u name in
+  let* prog_ir = desugar_unit imports u name in
   let prog_ir = ir_passes (!Flags.compile_mode) prog_ir name in
   phase "Interpreting" name;
   let open Interpret_ir in
   let flags = { trace = !Flags.trace; print_depth = !Flags.print_depth } in
-  interpret_prog flags prog_ir
+  Diag.return (interpret_prog flags prog_ir)
 
 let interpret_ir_files files =
-  Option.map
-    (fun (libs, progs, senv) -> interpret_ir_progs libs progs)
-    (Diag.flush_messages (load_progs parse_file files initial_stat_env))
+  Diag.flush_messages (Diag.bind
+    (load_progs parse_file files initial_stat_env)
+    (fun (libs, progs, _) -> interpret_ir_progs libs progs))